Harmonizing code that makes asset menu entries.
Working on removing MaterialManagerWrangler._make_asset_menu_entries().

The result will be to use Wrangler._make_asset_menu_entries() instead.

+++ scoremanager/wranglers/test/test_MaterialPackageWrangler_make_managermade_material_package.py
@@ -30,7 +30,8 @@ def test_MaterialPackageWrangler_make_managermade_material_package_01():
         manager = manager(path=path, session=session)
         assert manager._list() == directory_entries
     finally:
-        shutil.rmtree(path)
+        if os.path.exists(path):
+            shutil.rmtree(path)
     assert not os.path.exists(path)
