+++ scoremanager/wranglers/test/test_StylesheetWrangler_make_asset.py
@@ -14,11 +14,11 @@ def test_StylesheetWrangler_make_asset_01():
         )
 
     assert not os.path.exists(path)
-
     try:
         string = 'lmy new 1 test-stylesheet q'
         score_manager._run(pending_user_input=string, is_test=True)
         assert os.path.exists(path)
     finally:
-        os.remove(path)
-        assert not os.path.exists(path)
+        if os.path.exists(path):
+            os.remove(path)
+    assert not os.path.exists(path)
